[ENH] Change Rocket transformers base class to `BaseCollectionTransformer` (#637)
* rocket base class

* output convert

* tags and tests

* package rename

* rocket imports

* import

* docs
