+++ aeon/transformations/collection/convolution_based/_minirocket.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 """MiniRocket transformer."""
 
-__author__ = "angus924"
+__author__ = ["angus924"]
 __all__ = ["MiniRocket"]
 
 import multiprocessing
@@ -9,10 +9,10 @@ import multiprocessing
 import numpy as np
 from numba import get_num_threads, njit, prange, set_num_threads, vectorize
 
-from aeon.transformations.base import BaseTransformer
+from aeon.transformations.collection import BaseCollectionTransformer
 
 
-class MiniRocket(BaseTransformer):
+class MiniRocket(BaseCollectionTransformer):
     """MINImally RandOm Convolutional KErnel Transform (MiniRocket).
 
     MiniRocket [1]_ is an almost deterministic version of Rocket. If creates
@@ -47,7 +47,7 @@ class MiniRocket(BaseTransformer):
 
     Examples
     --------
-     >>> from aeon.transformations.collection.rocket import MiniRocket
+     >>> from aeon.transformations.collection.convolution_based import MiniRocket
      >>> from aeon.datasets import load_unit_test
      >>> X_train, y_train = load_unit_test(split="train")
      >>> X_test, y_test = load_unit_test(split="test")
@@ -62,9 +62,6 @@ class MiniRocket(BaseTransformer):
         "univariate-only": True,
         "fit_is_empty": False,
         "scitype:transform-output": "Primitives",
-        "scitype:instancewise": False,
-        "X_inner_mtype": "numpy3D",
-        "y_inner_mtype": "None",
     }
 
     def __init__(
@@ -79,7 +76,7 @@ class MiniRocket(BaseTransformer):
 
         self.n_jobs = n_jobs
         self.random_state = random_state
-        super(MiniRocket, self).__init__(_output_convert=False)
+        super(MiniRocket, self).__init__()
 
     def _fit(self, X, y=None):
         """Fits dilations and biases to input time series.
